refactor(s3): Clean up types, naming, and initialization
- Update docs to reference correct interfaces (ReadAtContentStore,
  IncrementalWriteStore instead of outdated StreamingContentStore)
- Change MaxParallelUploads and DeletionBatchSize from int to uint
- Rename WriteCache to Cache (field, config, and SetCache method)
- Rename statsCache to cachedStats for clarity
- Simplify S3 client endpoint initialization
- Clean up deletion queue initialization by extracting defaults

+++ pkg/store/content/s3/s3_incremental.go
@@ -148,7 +148,7 @@ func (s *S3ContentStore) FlushIncremental(ctx context.Context, id metadata.Conte
 	if maxConcurrent == 0 {
 		maxConcurrent = 4
 	}
-	if len(partsToUpload) > maxConcurrent {
+	if uint(len(partsToUpload)) > maxConcurrent {
 		partsToUpload = partsToUpload[:maxConcurrent]
 	}
